+++ indico/modules/events/registration/client/js/components/participant_list/ParticipantList.tsx
@@ -1,0 +1,13 @@
+export default function ParticipantListAccordion({
+  published,
+  totalParticipantCount,
+  tables,
+}: ParticipantListAccordionProps) {
+  return (
+    <ParticipantListAccordion
+      published={published}
+      totalParticipantCount={totalParticipantCount}
+      tables={tables}
+    />
+  );
+}
