+++ indico/modules/events/registration/client/js/components/participant_list/ParticipantListAccordion.module.scss
