+++ indico/modules/events/registration/client/js/components/participant_list/types.ts
@@ -1,0 +1,38 @@
+import {HTMLAttributes} from 'react';
+
+export interface TableColumnObj {
+  id: number;
+  text: string;
+  is_picture: boolean;
+}
+
+export interface TableRowObj {
+  id: number;
+  checked_in: boolean;
+  columns: TableColumnObj[];
+}
+
+export interface TableObj {
+  headers: string[];
+  rows: TableRowObj[];
+  num_participants: number;
+  show_checkin: boolean;
+  title?: string;
+}
+
+export interface ParticipantListAccordionProps {
+  totalParticipantCount: number;
+  published: boolean;
+  tables: TableObj[];
+}
+
+export interface AccordionParticipantsItemProps {
+  table: TableObj;
+  collapsible?: boolean;
+}
+
+export interface ParticipantCounterProps extends HTMLAttributes<HTMLSpanElement> {
+  totalCount: number;
+  hiddenCount: number;
+  styleName?: string;
+}
